requirements.txt
google-cloud-bigquery
python-dotenv
pandas
requests